+++ alma/views/js/alma-inpage.js
@@ -38,28 +38,29 @@ function removeLoaderButtonPayment(button) {
 }
 
 function onloadAlma() {
-    let radioButtons = document.querySelectorAll('input[name="payment-option"][data-module-name=alma]');
+    let radioButtons = document.querySelectorAll('input[name="payment-option"]');
 
     //Prestashop 1.7+
     radioButtons.forEach(function (input) {
         input.addEventListener("change", function () {
             let paymentOptionId = input.getAttribute('id');
             let blockForm = document.querySelector('#pay-with-' + paymentOptionId + '-form');
-            let formInpage = blockForm.querySelector('.alma-inpage');
             removeAlmaEventsFromPaymentButton();
             if (inPage !== undefined) {
                 inPage.unmount();
             }
-            if (this.dataset.moduleName === 'alma' && this.checked && formInpage) {
-                let installment = formInpage.dataset.installment;
-                if (installment === '1') {
-                    blockForm.hidden = true;
+            if (this.dataset.moduleName === 'alma') {
+                let formInpage = blockForm.querySelector('.alma-inpage');
+                if (this.checked && formInpage) {
+                    let installment = formInpage.dataset.installment;
+                    if (installment === '1') {
+                        blockForm.hidden = true;
+                    }
+                    let url = formInpage.dataset.action;
+
+                    inPage = createAlmaIframe(formInpage);
+                    mapPaymentButtonToAlmaPaymentCreation(url, inPage, input);
                 }
-                let url = formInpage.dataset.action;
-
-                inPage = createAlmaIframe(formInpage);
-
-                mapPaymentButtonToAlmaPaymentCreation(url, inPage, input);
             }
         });
     });