+++ alma/lib/Helpers/InsuranceHelper.php
@@ -24,10 +24,31 @@
 
 namespace Alma\PrestaShop\Helpers;
 
+use Alma\PrestaShop\Repositories\CartProductRepository;
+use Alma\PrestaShop\Repositories\ProductRepository;
 use PrestaShop\PrestaShop\Adapter\Shop\Context;
 
 class InsuranceHelper
 {
+    /**
+     * @var CartProductRepository
+     */
+    public $cartProductRepository;
+
+    /**
+     * @var ProductRepository
+     */
+    public $productRepository;
+
+    /**
+     *
+     */
+    public function __construct()
+    {
+        $this->cartProductRepository = new CartProductRepository();
+        $this->productRepository = new ProductRepository();
+    }
+
     /**
      * @return bool
      */
@@ -47,6 +68,26 @@ class InsuranceHelper
         return  (bool) version_compare(_PS_VERSION_, '1.6', '>=')
             && (bool) (int) SettingsHelper::get(ConstantsHelper::ALMA_ALLOW_INSURANCE, false)
             && (bool) (int) SettingsHelper::get(ConstantsHelper::ALMA_ACTIVATE_INSURANCE, false);
+    }
+
+    /**
+     * @return bool
+     */
+    public function hasInsuranceInCart()
+    {
+        $idInsuranceProduct = $this->productRepository->getProductIdByReference(ConstantsHelper::ALMA_INSURANCE_PRODUCT_REFERENCE);
+
+        if (!$idInsuranceProduct) {
+            return false;
+        }
+
+        /**
+         * @var \ContextCore $context
+         */
+        $context = \Context::getContext();
+        $idProduct = $this->cartProductRepository->hasProductInCart($idInsuranceProduct, $context->cart->id);
+
+        return (bool)$idProduct;
     }
 
 }