+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v3.1.2
+
+- hotfix: Select payment In-Page without open the modal and try to pay with other payment method
+- fix: Add loader payment button on Prestashop 1.6 before loading In-Page
+
 ## v3.1.1
 
 - hotfix: Pay now is broken if In-Page is disabled